Handle form error message pluralization
When submitting a form with invalid data, validators messages displayed in blocks "form_errors" and "field_errors" are not pluralized.

As a result, when applying a "MaxLength(255)" validator on a field, the following message is displayed :

"This value is too long. It should have 255 character or less.|This value is too long. It should have 255 characters or less."

This commit is very straightforward as it's a "copy" of https://github.com/symfony/TwigBridge/commit/e8f5968958b94f710a853854c886cd8fe9f107f2#Resources/views/Form/form_div_layout.html.twig