Fixed issue with widget help popovers not working on focus trigger
If the trigger for a widget help popover was set to 'focus' then the popover was not displayed. This was due to the fact that the data attributes are set on a surrounding div element rather than on the widget itself.

This change sets the 'data-selector' attribute to target the actual widget, by default. It is possible to override this and set a selector through the field (or global) configuration.

This change is a fix for issue: phiamo/MopaBootstrapBundle#793